+++ src/FootballCli/Model/FootballMatchPeriodScore.cs
@@ -1,5 +1,5 @@
 namespace Dr.FootballCli.Model;
 
 public readonly record struct FootballMatchPeriodScore(
-    int HomeTeam,
-    int AwayTeam);
+    int? Home,
+    int? Away);