+++ src/FootballCli/Model/FootballSeason.cs
@@ -4,5 +4,5 @@ public readonly record struct FootballSeason(
     int Id,
     DateTime StartDate,
     DateTime EndDate,
-    int CurrentMatchday,
+    int? CurrentMatchday,
     FootballSeasonWinner? Winner);