+++ src/FootballCli/Model/LeagueStanding.cs
@@ -6,11 +6,4 @@ public readonly record struct LeagueStanding(
     string Stage,
     string Type,
     string? Group,
-    List<LeaguePosition> Positions)
-{
-    public string PrettyPrintGroup =>
-        new CultureInfo("en-GB")
-            .TextInfo
-            .ToTitleCase(Group?.ToLower() ?? string.Empty)
-            .Replace('_', ' ');
-}
+    List<LeaguePosition> Positions);