+++ src/FootballCli/Program.cs
@@ -3,48 +3,42 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Dr.FootballCli;
 
-internal static class Program
+await Bootstrap().RunAsync(args);
+
+
+CommandApp Bootstrap()
 {
-    private static readonly string HostEnvironmentName = Environment.GetEnvironmentVariable("DOTNET_ENVIRONMENT") ?? "Production";
-
-    static async Task<int> Main(string[] args) =>
-        await Bootstrap().RunAsync(args)
+    var configuration = new ConfigurationBuilder()
+        // TOOD: This isn't always the execution folder.
+        .SetBasePath(Directory.GetCurrentDirectory())
+        .AddJsonFile($"AppSettings.json", optional: true)
+        .AddEnvironmentVariables("fbcli")
+        .Build()
     ;
 
+    var serviceCollection = new ServiceCollection()
+        .AddLogging(config => config.AddSimpleConsole())
+        .AddFootballCli(configuration)
+    ;
 
-    static CommandApp Bootstrap()
-    {
-        var configuration = new ConfigurationBuilder()
-            .SetBasePath(Directory.GetCurrentDirectory())
-            .AddJsonFile("appsettings.json", true, true)
-            .AddEnvironmentVariables("fbcli")
-            .Build()
-        ;
+    var registrar = new TypeRegistrar(serviceCollection);
+    var app = new CommandApp(registrar);
 
-        var serviceCollection = new ServiceCollection()
-            .AddLogging(config => config.AddSimpleConsole())
-            .AddFootballCli(configuration)
-        ;
+    app.Configure
+    (
+        config =>
+        {
+            config
+                .SetApplicationName("Football Cli")
+                .UseStrictParsing()
+                .AddCommands();
 
-        var registrar = new TypeRegistrar(serviceCollection);
-        var app = new CommandApp(registrar);
+            // Used during PRs to ensure help documentation is up to date.
+            if (Environment.GetEnvironmentVariable("FOOTBALL_CLI_VALIDATE_EXAMPLES") is not null)
+                config.ValidateExamples();
+        }
+    );
 
-        app.Configure
-        (
-            config =>
-            {
-                config
-                    .SetApplicationName("Football Cli")
-                    .UseStrictParsing()
-                    .UseEnvironmentSpecificConfig(HostEnvironmentName)
-                    .AddCommands()
-                ;
-            }
-        );
-
-
-        return app;
-    }
+    return app;
 }