+++ src/FootballCli/Startup.cs
@@ -1,9 +1,5 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Dr.FootballCli.Commands;
-using Dr.FootballCli.Config;
-using Dr.FootballCli.Repositories;
-using Dr.FootballCli.Views;
 
 namespace Dr.FootballCli;
 
@@ -12,51 +8,33 @@ public static class Startup
     public static IServiceCollection AddFootballCli(this IServiceCollection serviceCollection, IConfigurationRoot config)
     {
         serviceCollection
-            .Configure<SourceConfig>(config.GetSection("source"))
-            .Configure<FavouriteTeamConfig>(config.GetSection("favouriteTeam"))
-            .AddTransient<TableViewFactory>()
-            .AddTransient<CompetitionRepository>()
-            .AddTransient<LeagueRepository>()
-            .AddTransient<MatchRepository>()
-        ;
-
+            .Configure<ApiOptions>(config.GetSection("source"))
+            .AddSingleton<Client>();
 
         return serviceCollection;
     }
 
-    public static IConfigurator UseEnvironmentSpecificConfig(
-        this IConfigurator config,
-        string hostEnvironmentName
-    )
-    {
-        if(hostEnvironmentName.ToLower() == "development")
-            config
-                // smoke testing relies on ValidateExamples to ensure commands and examples are aligned
-                .ValidateExamples()
-                .PropagateExceptions()
-        ;
-
-
-        return config;
-    }
-
     public static IConfigurator AddCommands(this IConfigurator config)
     {
-        config.AddCommand<CompetitionCommand>("competition")
-            .WithDescription("View available competitions")
-            .WithExample(new[] { "competition" })
-        ;
+        config
+            .AddCommand<CompetitionsCommand>("competitions")
+            .WithDescription("Lists supported competitions.")
+            .WithAlias("competition")
+            .WithAlias("comps")
+            .WithAlias("comp")
+            .WithAlias("leagues")
+            .WithAlias("league")
+            .WithExample(new[] { "competitions" });
 
-        config.AddCommand<LiveScoresCommand>("live")
-            .WithDescription("View live scores")
-            .WithExample(new[] { "live", "-f", "--follow" })
-        ;
-
-        config.AddCommand<TableCommand>("table")
-            .WithDescription("View league standings")
-            .WithExample(new[] { "table", "-f", "--full-table" })
-        ;
-
+        config
+            .AddCommand<MatchesCommand>("matches")
+            .WithDescription("Lists supported competitions.")
+            .WithAlias("fixtures")
+            .WithAlias("fixture")
+            .WithAlias("results")
+            .WithAlias("result")
+            .WithAlias("match")
+            .WithExample(new[] { "matches", "pl", "--from", "-1" });
 
         return config;
     }